requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.10.3
djangorestframework-csv==2.1.0
factory-boy==2.12.0
Faker==3.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
unicodecsv==0.14.1
